+++ tests/testthat/test-semmcci-mc-simple-med-defined-inequality.R
@@ -3,17 +3,15 @@ lapply(
   X = 1,
   FUN = function(i,
                  n,
+                 R,
+                 tol,
                  text) {
-    R <- 10L
     message(text)
-    seed <- sample.int(
-      n = .Machine$integer.max,
-      size = 1
-    )
-    set.seed(42) # seed for data generation
-    cp <- 0
-    b <- 0.50
-    a <- 0.50
+    seed <- 42
+    set.seed(seed)
+    cp <- 0.00
+    b <- 0.10
+    a <- 0.10
     sigma2ey <- 1 - b^2 - cp^2 - 2 * a * b * cp
     sigma2em <- 1 - a^2
     sigma2x <- 1
@@ -52,22 +50,20 @@ lapply(
       text,
       {
         testthat::expect_equal(
-          results$thetahatstar[, 1:7],
-          answers
-        )
-        testthat::expect_equal(
           results$thetahat$est[1:7],
           lavaan::parameterEstimates(fit)$est,
           check.attributes = FALSE
         )
-        testthat::expect_equal(
-          results$ci["ab", "0.05%"],
-          quantile(answers[, "ab"], .0005),
-          check.attributes = FALSE
+        testthat::expect_true(
+          abs(
+            results$ci["ab", "0.05%"] - quantile(answers[, "ab"], .0005)
+          ) <= tol
         )
       }
     )
   },
-  n = 100L,
+  n = 1000L,
+  R = 1000L,
+  tol = 0.05,
   text = "test-semmcci-mc-simple-med-defined-inequality"
 )