+++ tests/testthat/test-semmcci-mc-simple-med-std-defined.R
@@ -5,18 +5,11 @@ lapply(
                  n,
                  text) {
     message(text)
-    seed <- sample.int(
-      n = .Machine$integer.max,
-      size = 1
-    )
-    coefs <- stats::runif(
-      n = 3,
-      min = 0.0,
-      max = 0.5
-    )
-    cp <- coefs[1]
-    b <- coefs[2]
-    a <- coefs[3]
+    seed <- 42
+    set.seed(seed)
+    cp <- 0.00
+    b <- 0.10
+    a <- 0.10
     sigma2ey <- 1 - b^2 - cp^2 - 2 * a * b * cp
     sigma2em <- 1 - a^2
     sigma2x <- 1
@@ -58,18 +51,18 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          results$thetahatstar.std[3, ],
+          results$thetahatstar_std[3, ],
           lavaan::standardizedSolution(fit)$est.std,
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          results$ci.std["ab", "0.05%"],
-          quantile(results$thetahatstar.std[, "ab"], .0005),
+          results$ci_std["ab", "0.05%"],
+          quantile(results$thetahatstar_std[, "ab"], .0005),
           check.attributes = FALSE
         )
       }
     )
   },
-  n = 100L,
+  n = 1000L,
   text = "test-semmcci-mc-simple-med-std-defined"
 )